OneDrive/Documents/cv/backend/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
opencv-python==4.7.0.72
numpy==1.24.3
Pillow==9.5.0
